+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed `xplatform.loadFileUTF8` when trying to open nonexistent file (#454)
 - Bugfix: fix an incorrect check in the recovery router code which might lead to
   the state cell-pool being released prematurely (#446)
+- Allocating SLH for http server with configurable value 'httpRequestHeapMaxBlocks' in yaml (#447).
 
 ## `2.16.0`
 - No yaml value converted to null (#442)